polish Home screen UI with global theme, modal editing, and swipe actions
