+++ app/(tabs)/add.tsx
@@ -1,6 +1,6 @@
 // app/(tabs)/add.tsx
 import React from 'react';
-import AddItemScreen from '../../screens/AddItemScreen';
+import AddItemScreen from '../../src/screens/AddItemScreen';
 
 export default function Add() {
   return <AddItemScreen />;
