+++ app/(tabs)/index.tsx
@@ -1,6 +1,6 @@
 // app/(tabs)/index.tsx
 import React from 'react';
-import HomeScreen from '../../screens/HomeScreen';
+import HomeScreen from '../../src/screens/HomeScreen';
 
 export default function Home() {
   return <HomeScreen />;
