+++ theme/colors.ts
@@ -1,0 +1,16 @@
+export const colors = {
+  primary: '#007AFF',
+  background: '#ffffff',
+  cardBackground: '#f9f9f9',
+
+  textPrimary: '#1c1c1e',       // ✅ Strong readable text color (titles)
+  textSecondary: '#6e6e73',     // ✅ Muted text (subtitles, hints)
+
+  inputBackground: '#f2f2f7',
+  inputBorder: '#ddd',
+  border: '#d1d1d6',
+
+  danger: '#ff3b30',
+  success: '#34c759',
+  shadow: 'rgba(0, 0, 0, 0.1)',  
+};
