+++ theme/index.ts
@@ -1,0 +1,3 @@
+export * from './colors';
+export * from './typography';
+
