+++ theme/metrics.ts
@@ -1,0 +1,8 @@
+// theme/metrics.ts
+export const Spacing = {
+  xs: 4,
+  sm: 8,
+  md: 16,
+  lg: 24,
+  xl: 32,
+};
