+++ theme/typography.ts
@@ -1,0 +1,7 @@
+export const typography = {
+  heading: { fontSize: 22, fontWeight: 'bold' as const },
+  body: { fontSize: 16, color: '#333' },
+  small: { fontSize: 12, color: '#888' },
+  title: { fontSize: 22, fontWeight: '600' as const, color: '#000' },
+  label: { fontSize: 13, fontWeight: '500' as const, color: '#555' },
+};
